+++ app/src/main/java/com/bnhc/opengl/model/Point.java
@@ -9,7 +9,7 @@ public class Point {
     public float y;
     public float z;
 
-    public Point(int x, int y, int z) {
+    public Point(float x, float y, float z) {
         this.x = x;
         this.y = y;
         this.z = z;
